経験値関連の調整

+++ Assets/Scripts/LootPanel.cs
@@ -297,7 +297,7 @@ public class LootPanel : MonoBehaviour
         }
         //Sort();
     }
-    public void AddExp(int amount)
+    public void AddExp(float amount)
     {
         expOrbs += expeditionManager.GetExpAmount(amount);
     }
